fixed: fehlerhafte typenüberführung
double -> float-Überführung in Artikel::getPreis-Funktionen führte zu Genauigkeitsfehlern in der Repräsentation der double-Werte normalpreis und angebotspreis.

+++ source/artikel.cpp
@@ -19,12 +19,12 @@ Artikel::Artikel()
     naechste = NULL;
 }
 
-float Artikel::getPreis()
+double Artikel::getPreis()
 {
     if (angebotspreis != NULL) return angebotspreis;
     return normalpreis;
 }
-float Artikel::getPreis(int i)
+double Artikel::getPreis(int i)
 {
     if (i == 0) return normalpreis;
     if (i == 1) return angebotspreis;
